Move ServiceNames to separate file.
