+++ server/data-access/index.ts
@@ -1,12 +1,10 @@
 import { AxiosRequestConfig, AxiosRequestHeaders } from 'axios';
 import { IncomingHttpHeaders } from 'http';
 // import * as getAutomationAnalyticsDescriptor from './automation-analytics';
-import prepareServiceCall, {
-  ServiceCallFunction,
-  ServiceNames,
-} from './call-service';
+import prepareServiceCall, { ServiceCallFunction } from './call-service';
 import demoDescriptor from './demoDescriptor';
 import complianceDescriptor from './complianceDescriptor';
+import ServiceNames from './service-names';
 
 const templateMapper: {
   [key in ServiceNames]: ServiceCallFunction;
