+++ server/data-access/service-names.ts
@@ -1,0 +1,6 @@
+enum ServiceNames {
+  'compliance' = 'compliance',
+  'demo' = 'demo',
+}
+
+export default ServiceNames;
